requirements.txt
alphabet-detector==0.0.7
asn1crypto==1.2.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.1.7
django-admin-tools==0.8.1
django-admin-tools-stats==0.9.0
django-bower==5.2.0
django-colorfield==0.1.15
django-crequest==2018.5.11
django-db-mailer==2.4.0
django-extra-fields==1.2.0
django-filter==2.2.0
django-jet==1.0.8
django-jsonfield==1.4.0
django-localflavor==2.2
django-model-utils==3.2.0
django-modeltranslation==0.14
django-nvd3==0.9.7
django-object-tool==0.0.9
django-qsstats-magic==1.1.0
django-solo==1.1.3
djangorestframework==3.10.3
djcacheutils==3.0.0
idna==2.8
Jinja2==2.10.3
jsonfield==2.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
optionaldict==0.1.1
Pillow==6.2.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
python-social-auth==0.3.6
python-stdnum==1.11
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-core==3.2.0
sqlparse==0.3.0
twython==3.7.0
Unidecode==1.1.1
urllib3==1.25.6
wechat-django==0.3.3
wechatpy==1.8.5
xmltodict==0.12.0
rest-framework-generic-relations